-YuiErrorMsg changed to accommodate the new error handling system -Changed errors in yabause.c to use new error handling -Moved SDL specific stuff from yabause.c to yui -Changed a few variable names -Fixed at least one bug

+++ branches/yabauserewrite/yabause/src/yabause.h
@@ -36,8 +36,8 @@ typedef struct
    int DecilineStop;
    u32 Duf;
    u32 CycleCountII;
-   int IsSSH2Running;
-   int IsM68KRunning;
+   u8 IsSSH2Running;
+   u8 IsM68KRunning;
 } yabsys_struct;
 
 extern yabsys_struct yabsys;
